requirements.txt
python-dotenv==1.0.1
langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.2.14
langchain-ollama==0.2.2
openai==1.59.3 
google-search-results==2.4.2
flask==3.1.0 
tweepy==4.14.0 
black==24.10.0